+++ examples/crd_openapi.rs
@@ -25,10 +25,11 @@ pub struct FooStatus {
 
 type Foo = Object<FooSpec, FooStatus>;
 
-fn main() -> Result<(), failure::Error> {
+#[tokio::main]
+async fn main() -> Result<(), failure::Error> {
     std::env::set_var("RUST_LOG", "info,kube=trace");
     env_logger::init();
-    let config = config::load_kube_config().expect("failed to load kubeconfig");
+    let config = config::load_kube_config().await?;
     let client = APIClient::new(config);
 
     // Manage the CRD
@@ -37,7 +38,7 @@ fn main() -> Result<(), failure::Error> {
     // Delete any old versions of it first:
     let dp = DeleteParams::default();
     // but ignore delete err if not exists
-    let _ = crds.delete("foos.clux.dev", &dp).map(|res| {
+    let _ = crds.delete("foos.clux.dev", &dp).await.map(|res| {
         res.map_left(|o| {
             info!("Deleted {}: ({:?})", o.metadata.name,
                 o.status.unwrap().conditions.unwrap().last());
@@ -77,7 +78,7 @@ fn main() -> Result<(), failure::Error> {
     info!("Creating CRD foos.clux.dev");
     let pp = PostParams::default();
     let patch_params = PatchParams::default();
-    match crds.create(&pp, serde_json::to_vec(&foocrd)?) {
+    match crds.create(&pp, serde_json::to_vec(&foocrd)?).await {
         Ok(o) => {
             info!("Created {} ({:?})", o.metadata.name, o.status);
             debug!("Created CRD: {:?}", o.spec);
@@ -105,13 +106,13 @@ fn main() -> Result<(), failure::Error> {
         "metadata": { "name": "baz" },
         "spec": { "name": "baz", "info": "old baz", "replicas": 1 },
     });
-    let o = foos.create(&pp, serde_json::to_vec(&f1)?)?;
+    let o = foos.create(&pp, serde_json::to_vec(&f1)?).await?;
     assert_eq!(f1["metadata"]["name"], o.metadata.name);
     info!("Created {}", o.metadata.name);
 
     // Verify we can get it
     info!("Get Foo baz");
-    let f1cpy = foos.get("baz")?;
+    let f1cpy = foos.get("baz").await?;
     assert_eq!(f1cpy.spec.info, "old baz");
 
     // Replace its spec
@@ -126,13 +127,13 @@ fn main() -> Result<(), failure::Error> {
         },
         "spec": { "name": "baz", "info": "new baz", "replicas": 1 },
     });
-    let f1_replaced = foos.replace("baz", &pp, serde_json::to_vec(&foo_replace)?)?;
+    let f1_replaced = foos.replace("baz", &pp, serde_json::to_vec(&foo_replace)?).await?;
     assert_eq!(f1_replaced.spec.name, "baz");
     assert_eq!(f1_replaced.spec.info, "new baz");
     assert!(f1_replaced.status.is_none());
 
     // Delete it
-    foos.delete("baz", &dp)?.map_left(|f1del| {
+    foos.delete("baz", &dp).await?.map_left(|f1del| {
         assert_eq!(f1del.spec.info, "old baz");
     });
 
@@ -146,7 +147,7 @@ fn main() -> Result<(), failure::Error> {
         "spec": FooSpec { name: "qux".into(), replicas: 0, info: "unpatched qux".into() },
         "status": FooStatus::default(),
     });
-    let o = foos.create(&pp, serde_json::to_vec(&f2)?)?;
+    let o = foos.create(&pp, serde_json::to_vec(&f2)?).await?;
     info!("Created {}", o.metadata.name);
 
     // Update status on qux
@@ -161,7 +162,7 @@ fn main() -> Result<(), failure::Error> {
         },
         "status": FooStatus { is_bad: true, replicas: 0 }
     });
-    let o = foos.replace_status("qux", &pp, serde_json::to_vec(&fs)?)?;
+    let o = foos.replace_status("qux", &pp, serde_json::to_vec(&fs)?).await?;
     info!("Replaced status {:?} for {}", o.status, o.metadata.name);
     assert!(o.status.unwrap().is_bad);
 
@@ -169,18 +170,18 @@ fn main() -> Result<(), failure::Error> {
     let fs = json!({
         "status": FooStatus { is_bad: false, replicas: 1 }
     });
-    let o = foos.patch_status("qux", &patch_params, serde_json::to_vec(&fs)?)?;
+    let o = foos.patch_status("qux", &patch_params, serde_json::to_vec(&fs)?).await?;
     info!("Patched status {:?} for {}", o.status, o.metadata.name);
     assert!(!o.status.unwrap().is_bad);
 
     info!("Get Status on Foo instance qux");
-    let o = foos.get_status("qux")?;
+    let o = foos.get_status("qux").await?;
     info!("Got status {:?} for {}", o.status, o.metadata.name);
     assert!(!o.status.unwrap().is_bad);
 
     // Check scale subresource:
     info!("Get Scale on Foo instance qux");
-    let scale = foos.get_scale("qux")?;
+    let scale = foos.get_scale("qux").await?;
     info!("Got scale {:?} - {:?}", scale.spec, scale.status);
     assert_eq!(scale.status.unwrap().replicas, 1);
 
@@ -188,7 +189,7 @@ fn main() -> Result<(), failure::Error> {
     let fs = json!({
         "spec": { "replicas": 2 }
     });
-    let o = foos.patch_scale("qux", &patch_params, serde_json::to_vec(&fs)?)?;
+    let o = foos.patch_scale("qux", &patch_params, serde_json::to_vec(&fs)?).await?;
     info!("Patched scale {:?} for {}", o.spec, o.metadata.name);
     assert_eq!(o.status.unwrap().replicas, 1);
     assert_eq!(o.spec.replicas.unwrap(), 2); // we only asked for more
@@ -198,21 +199,21 @@ fn main() -> Result<(), failure::Error> {
     let patch = json!({
         "spec": { "info": "patched qux" }
     });
-    let o = foos.patch("qux", &patch_params, serde_json::to_vec(&patch)?)?;
+    let o = foos.patch("qux", &patch_params, serde_json::to_vec(&patch)?).await?;
     info!("Patched {} with new name: {}", o.metadata.name, o.spec.name);
     assert_eq!(o.spec.info, "patched qux");
     assert_eq!(o.spec.name, "qux"); // didn't blat existing params
 
     // Check we have 1 remaining instance
     let lp = ListParams::default();
-    let res = foos.list(&lp)?;
+    let res = foos.list(&lp).await?;
     assert_eq!(res.items.len(), 1);
 
     // Delete the last - expect a status back (instant delete)
-    assert!(foos.delete("qux", &dp)?.is_right());
+    assert!(foos.delete("qux", &dp).await?.is_right());
 
     // Cleanup the full collection - expect a wait
-    match foos.delete_collection(&lp)? {
+    match foos.delete_collection(&lp).await? {
         Left(list) => {
             let deleted = list.items.into_iter().map(|i| i.metadata.name).collect::<Vec<_>>();
             info!("Deleted collection of foos: {:?}", deleted);