+++ examples/log_openapi.rs
@@ -1,16 +1,17 @@
 #[macro_use] extern crate log;
 use failure::{err_msg, Error};
 use kube::{
-    api::{Api, Log, LogParams},
+    api::{Api, LogParams},
     client::APIClient,
     config,
 };
 use std::env;
 
-fn main() -> Result<(), Error> {
+#[tokio::main]
+async fn main() -> Result<(), Error> {
     std::env::set_var("RUST_LOG", "info,kube=trace");
     env_logger::init();
-    let config = config::load_kube_config().expect("failed to load kubeconfig");
+    let config = config::load_kube_config().await?;
     let client = APIClient::new(config);
     let namespace = std::env::var("NAMESPACE").unwrap_or("default".into());
 
@@ -22,7 +23,7 @@ fn main() -> Result<(), Error> {
     // because we don't specify lp.container the pod must have only 1 container
     let pods = Api::v1Pod(client).within(&namespace);
     let lp = LogParams::default();
-    let plog = pods.log(&mypod, &lp)?;
+    let plog = pods.log(&mypod, &lp).await?;
     info!("Got pod {} log: {}", &mypod, &plog);
 
     Ok(())