+++ src/api/mod.rs
@@ -31,7 +31,6 @@ pub use typed::{
     Scale,
     ScaleSpec,
     ScaleStatus,
-    Log
 };
 
 mod resource;