+++ src/config/mod.rs
@@ -44,16 +44,8 @@ impl Configuration {
 }
 
 /// Returns a config includes authentication and cluster infomation from kubeconfig file.
-///
-/// # Example
-/// ```no_run
-/// use kube::config;
-///
-/// let kubeconfig = config::load_kube_config()
-///     .expect("failed to load kubeconfig");
-/// ```
-pub fn load_kube_config() -> Result<Configuration> {
-    load_kube_config_with(Default::default())
+pub async fn load_kube_config() -> Result<Configuration> {
+    load_kube_config_with(Default::default()).await
 }
 
 /// ConfigOptions stores options used when loading kubeconfig file.
@@ -65,17 +57,9 @@ pub struct ConfigOptions {
 }
 
 /// Returns a config which includes authentication and cluster information from kubeconfig file.
-///
-/// # Example
-/// ```no_run
-/// use kube::config;
-///
-/// let kubeconfig = config::load_kube_config()
-///     .expect("failed to load kubeconfig");
-/// ```
-pub fn load_kube_config_with(options: ConfigOptions) -> Result<Configuration> {
+pub async fn load_kube_config_with(options: ConfigOptions) -> Result<Configuration> {
 
-    let result = create_client_builder(options)?;
+    let result = create_client_builder(options).await?;
 
     Ok(Configuration::new(
         result.1.cluster.server,
@@ -87,22 +71,12 @@ pub fn load_kube_config_with(options: ConfigOptions) -> Result<Configuration> {
 /// Returns a client builder and config loader, based on the cluster information from the kubeconfig file.
 ///
 /// This allows to create your custom reqwest client for using with the cluster API.
-///
-/// # Example
-/// ```no_run
-/// use kube::config;
-///
-/// let client_builder_result = config::create_client_builder(Default::default())
-///     .expect("failed to load kubeconfig");
-/// let client_builder = client_builder_result.0;
-/// let loader = client_builder_result.1;
-/// ```
-pub fn create_client_builder(options: ConfigOptions) -> Result<(ClientBuilder,KubeConfigLoader)> {
+pub async fn create_client_builder(options: ConfigOptions) -> Result<(ClientBuilder,KubeConfigLoader)> {
     let kubeconfig = utils::find_kubeconfig()
         .context(ErrorKind::KubeConfig("Unable to load file".into()))?;
 
     let loader =
-        KubeConfigLoader::load(kubeconfig, options.context, options.cluster, options.user)?;
+        KubeConfigLoader::load(kubeconfig, options.context, options.cluster, options.user).await?;
 
     let token = match &loader.user.token {
         Some(token) => Some(token.clone()),
@@ -171,15 +145,8 @@ pub fn create_client_builder(options: ConfigOptions) -> Result<(ClientBuilder,Ku
 }
 
 /// Returns a config which is used by clients within pods on kubernetes.
+///
 /// It will return an error if called from out of kubernetes cluster.
-///
-/// # Example
-/// ```no_run
-/// use kube::config;
-///
-/// let kubeconfig = config::incluster_config()
-///     .expect("failed to load incluster config");
-/// ```
 pub fn incluster_config() -> Result<Configuration> {
     let server = incluster_config::kube_server().ok_or_else(||
         Error::from(ErrorKind::KubeConfig(format!(