+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.18.0 / 2019-XX-XX
+==================
+  * api converted to use async/await with 1.39.0
+  * `Log` trait removed in favour of internal marker trait
+
 0.17.0 / 2019-10-22
 ==================
   * Add support for oidc providerss with `auth-provider` w/o `access-token` - #70